Move initialize_tf_variables under softlearning.utils.tensorflow
